Ian Isely - 2/20/21
Starting with the main window, I changed the button layout slightly to
make the buttons more distinguishable from each other, and added "Add
Track" and "Remove Track" buttons. I moved the "Information" button to
the bottom of the window so it is less in the user's way.

I implemented the information window that will open upon clicking the
"Information" button. This window contains the product description,
developer names, and a message box containing information about Qt. The
code for this new window is contained in the infowindow.cpp,
infowindow.h, and infowindow.ui files which will be added with this
commit.