Ian Isely - 3/9/21
I added functionality to the "Add Tracks" and "Remove Tracks" buttons
allowing the user to add/remove one or more .mp3 files to a "Songs"
folder that is created by the program in the C:\Users\usr\Music
directory. The ui was slightly updated to accomodate the multi file
functionality.