Ian Isely - 3/22/21
I improved the list widget and play/pause button functionalities. The
color of the playing/paused track now persists after adding or removing
tracks, and a safeguard is in place if the user tries to remove the
currently playing song. The "songs" resource folder was deleted since it
is no longer in use.